+++ packages/components/src/DataList/DataList.test.tsx
@@ -564,7 +564,7 @@ describe("DataList", () => {
       ).not.toBeInTheDocument();
     });
 
-    it("should call the scrollIntoView on the target element", () => {
+    it("should call the scrollIntoView on the target element", async () => {
       const scrollIntoViewMock = jest.fn();
       window.HTMLElement.prototype.scrollIntoView = scrollIntoViewMock;
       render(
@@ -580,7 +580,9 @@ describe("DataList", () => {
 
       expect(scrollIntoViewMock).not.toHaveBeenCalled();
 
-      userEvent.click(screen.getByRole("button", { name: "Back to top" }));
+      await userEvent.click(
+        screen.getByRole("button", { name: "Back to top" }),
+      );
       expect(scrollIntoViewMock).toHaveBeenCalledWith({
         behavior: "smooth",
         block: "nearest",