+++ packages/components/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.53.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.52.1...@jobber/components@4.53.0) (2023-10-23)
+
+
+### Features
+
+* **components:** JOB-80773 Datatable scroll to top pagination loading ([#1595](https://github.com/GetJobber/atlantis/issues/1595)) ([e7dfc3c](https://github.com/GetJobber/atlantis/commit/e7dfc3c584691cc9667627f6eb59d49b48749f1d))
+
+
+
+
+
+## [4.52.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.52.0...@jobber/components@4.52.1) (2023-10-23)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
+# [4.52.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.51.0...@jobber/components@4.52.0) (2023-10-20)
+
+
+### Features
+
+* **components:** Tabs supports caching active tab between page loads ([#1591](https://github.com/GetJobber/atlantis/issues/1591)) ([4bc6f8e](https://github.com/GetJobber/atlantis/commit/4bc6f8e6f9ebf2922e0fbd12df85b43966d8ba34))
+
+
+
+
+
+# [4.51.0](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.50.2...@jobber/components@4.51.0) (2023-10-19)
+
+
+### Features
+
+* **components:** Combobox Header ([#1570](https://github.com/GetJobber/atlantis/issues/1570)) ([ec41d1c](https://github.com/GetJobber/atlantis/commit/ec41d1c996a0bac4408dea0a3bdc9b8a95200428))
+
+
+
+
+
+## [4.50.2](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.50.1...@jobber/components@4.50.2) (2023-10-19)
+
+**Note:** Version bump only for package @jobber/components
+
+
+
+
+
 ## [4.50.1](https://github.com/GetJobber/atlantis/compare/@jobber/components@4.50.0...@jobber/components@4.50.1) (2023-10-19)
 
 **Note:** Version bump only for package @jobber/components